requirements.txt
appdirs==1.4.0
certifi==2017.1.23
future==0.16.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
python-telegram-bot==5.3.0
pytz==2016.10
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
urllib3==1.20
